+++ CHANGES.rst
@@ -16,10 +16,13 @@ Changes:
   requirement classes for ``weaver:BuiltinRequirement``, ``weaver:WPS1Requirement``, ``weaver:OGCAPIRequirement``
   and ``weaver:ESGF-CWTRequirement`` to avoid missing reference warnings that were previously raised by ``cwltool``
   due to `Application Packages` using their non-``weaver`` namespaced classes in ``hints``.
+- Add ``weaver.wps_client_headers_filter`` setting that allows filtering of specific `WPS` request headers from the
+  incoming request to be passed down to the `WPS` client employed to interact with the `WPS` provider
+  (fixes `#600 <https://github.com/crim-ca/weaver/issues/600>`_).
 
 Fixes:
 ------
-- No change.
+- Fix ``moto>=5`` used in tests to mock AWS S3 operations that replaced ``mock_s3`` context manager by ``mock_aws``.
 
 .. _changes_5.0.0: